moved SelfTestDEsc and created some javadoc
