+++ pepper-framework/src/test/java/org/corpus_tools/pepper/core/ModuleFitnessCheckerTest.java
@@ -16,11 +16,11 @@ import org.corpus_tools.pepper.common.ModuleFitness.Fitness;
 import org.corpus_tools.pepper.common.ModuleFitness.FitnessFeature;
 import org.corpus_tools.pepper.common.Pepper;
 import org.corpus_tools.pepper.exceptions.PepperFWException;
-import org.corpus_tools.pepper.impl.SelfTestDesc;
 import org.corpus_tools.pepper.modules.PepperExporter;
 import org.corpus_tools.pepper.modules.PepperImporter;
 import org.corpus_tools.pepper.modules.PepperManipulator;
 import org.corpus_tools.pepper.modules.PepperModule;
+import org.corpus_tools.pepper.modules.SelfTestDesc;
 import org.corpus_tools.pepper.modules.coreModules.DoNothingImporter;
 import org.corpus_tools.pepper.modules.coreModules.DoNothingManipulator;
 import org.corpus_tools.pepper.testFramework.PepperTestUtil;
@@ -300,7 +300,7 @@ public class ModuleFitnessCheckerTest {
 		final Pepper pepper = PepperTestUtil.createDefaultPepper();
 		final SelfTestDesc desc = mock(SelfTestDesc.class);
 		when(desc.getInputCorpusPath()).thenReturn(URI.createFileURI(PepperTestUtil.getTestResources() + "selfTest/sampleCorpus/in/"));
-		when(desc.getOutputCorpusPath()).thenReturn(URI.createFileURI(PepperTestUtil.getTestResources() + "selfTest/sampleCorpus/out/"));
+		when(desc.getExpectedCorpusPath()).thenReturn(URI.createFileURI(PepperTestUtil.getTestResources() + "selfTest/sampleCorpus/out/"));
 		when(desc.isValid(Matchers.anyListOf(String.class))).thenReturn(true);
 		when(desc.compare(any(SaltProject.class), any(SaltProject.class))).thenReturn(true);
 		final PepperImporter importer = spy(DoNothingImporter.class);
@@ -321,7 +321,7 @@ public class ModuleFitnessCheckerTest {
 		final Pepper pepper = PepperTestUtil.createDefaultPepper();
 		final SelfTestDesc desc = mock(SelfTestDesc.class);
 		when(desc.getInputCorpusPath()).thenReturn(URI.createFileURI(PepperTestUtil.getTestResources() + "selfTest/sampleCorpus/in/"));
-		when(desc.getOutputCorpusPath()).thenReturn(URI.createFileURI(PepperTestUtil.getTestResources() + "selfTest/sampleCorpus/out/"));
+		when(desc.getExpectedCorpusPath()).thenReturn(URI.createFileURI(PepperTestUtil.getTestResources() + "selfTest/sampleCorpus/out/"));
 		when(desc.isValid(Matchers.anyListOf(String.class))).thenReturn(true);
 		when(desc.compare(any(SaltProject.class), any(SaltProject.class))).thenReturn(false);
 		final PepperImporter importer = spy(DoNothingImporter.class);
@@ -342,7 +342,7 @@ public class ModuleFitnessCheckerTest {
 		final Pepper pepper = PepperTestUtil.createDefaultPepper();
 		final SelfTestDesc desc = mock(SelfTestDesc.class);
 		when(desc.getInputCorpusPath()).thenReturn(URI.createFileURI(PepperTestUtil.getTestResources() + "selfTest/sampleCorpus/in/"));
-		when(desc.getOutputCorpusPath()).thenReturn(URI.createFileURI(PepperTestUtil.getTestResources() + "selfTest/sampleCorpus/out/"));
+		when(desc.getExpectedCorpusPath()).thenReturn(URI.createFileURI(PepperTestUtil.getTestResources() + "selfTest/sampleCorpus/out/"));
 		when(desc.isValid(Matchers.anyListOf(String.class))).thenReturn(true);
 		when(desc.compare(any(SaltProject.class), any(SaltProject.class))).thenReturn(false);
 		final PepperManipulator importer = spy(DoNothingManipulator.class);
